+++ app/src/main/java/com/github/droidworksstudio/launcher/ui/drawer/DrawFragment.kt
@@ -160,12 +160,12 @@ class DrawFragment : Fragment(),
         return object : OnSwipeTouchListener(context) {
             override fun onSwipeLeft() {
                 super.onSwipeLeft()
-                findNavController().popBackStack()
+                findNavController().navigateUp()
             }
 
             override fun onSwipeRight() {
                 super.onSwipeRight()
-                findNavController().popBackStack()
+                findNavController().navigateUp()
             }
         }
     }